analyser_service/pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "analyser_service"
version = "0.1.0"
description = "CAD analyser service with FreeCAD for extracting geometry statistics"
authors = [{name="Michel Kruse"}]
dependencies = [
    "fastapi",
    "uvicorn",
    "requests",
    "numpy",
    "pandas",
    "python-multipart",
]

[tool.setuptools]
package-dir = { "" = "src" }

[tool.setuptools.packages.find]
where = ["src"]
